requirements/base.txt
crispy-bootstrap5==0.7
Django==4.1.6
django-allauth==0.52.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
python-dotenv==0.21.1
readtime==2.0.0
